OSS 042023

+++ shell/static_plugins/comp_surf/comp_surf.cc
@@ -15,10 +15,8 @@
 #include "comp_surf.h"
 
 #include <dlfcn.h>
-#include <flutter/fml/logging.h>
 #include <flutter/standard_method_codec.h>
 
-#include "../../view/flutter_view.h"
 #include "engine.h"
 
 void CompositorSurfacePlugin::OnPlatformMessage(
@@ -159,7 +157,7 @@ void CompositorSurfacePlugin::OnPlatformMessage(
           view->CreateSurface(h_module, assets_path, cache_folder, misc_folder,
                               type, z_order, sync, width, height, x, y);
 
-      auto context = view->GetSurfaceContext(index);
+      auto context = view->GetSurfaceContext(static_cast<int64_t>(index));
 
       auto value = flutter::EncodableValue(flutter::EncodableMap{
           {flutter::EncodableValue("result"), flutter::EncodableValue(0)},
